requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
pymongo==4.5.0
python-dotenv==1.0.0